ICUAS READY

+++ casa_px4_mocap/src/casa_flight_mode.cpp
@@ -10,7 +10,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "px4_msgs/msg/vehicle_odometry.hpp"
 #include "geometry_msgs/msg/pose_stamped.hpp"
-#include "std_msgs/msg/int64.hpp"
+#include "std_msgs/msg/int32.hpp"
 
 using namespace std::chrono_literals;
 px4_msgs::msg::VehicleOdometry::_timestamp_type dbk_timestamp, dbk_timestamp_sample;
@@ -24,8 +24,8 @@ public:
     flightMode()
             : Node("rrc_flight_mode"), count_(0)
     {
-        mode_publisher_ = this->create_publisher<std_msgs::msg::Int64>("/fmu/flight_mode",10);
-//        uav_publisher_ = this->create_publisher<std_msgs::msg::Int64>("/fmu/uav_number",10);
+        mode_publisher_ = this->create_publisher<std_msgs::msg::Int32>("/fmu/flight_mode",10);
+//        uav_publisher_ = this->create_publisher<std_msgs::msg::Int32>("/fmu/uav_number",10);
         timer_ = this->create_wall_timer(10ms, std::bind(&flightMode::timer_callback,this));
     }
 
@@ -34,7 +34,7 @@ private:
 
     void timer_callback()
     {
-        auto flight_mode_msg = std_msgs::msg::Int64();
+        auto flight_mode_msg = std_msgs::msg::Int32();
 //        auto uav_number_msg = std_msgs::msg::Int64 ();
 
 //        std::cout<<"Enter the UAV number"<<std::endl;
@@ -51,7 +51,7 @@ private:
     }
 
     rclcpp::TimerBase::SharedPtr timer_;
-    rclcpp::Publisher<std_msgs::msg::Int64>::SharedPtr mode_publisher_;
+    rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr mode_publisher_;
 //    rclcpp::Publisher<std_msgs::msg::Int64>::SharedPtr uav_publisher_;
     size_t count_;
     int flight_mode;
